+++ src/modules/dgt.power.codegeneration/Base/TypescriptGeneratorVersion.cs
@@ -1,0 +1,10 @@
+// Copyright (c) DIGITALL Nature. All rights reserved
+// DIGITALL Nature licenses this file to you under the Microsoft Public License.
+
+namespace dgt.power.codegeneration.Base;
+
+public enum TypescriptGeneratorVersion
+{
+    Light,
+    Full
+}