+++ src/modules/dgt.power.codegeneration/Model/FormDetail.cs
@@ -12,7 +12,18 @@ public class FormDetail
         Grids = new HashSet<string>();
     }
 
+    public Dictionary<string, TabDetail> TabDetails { get; } = [];
     public Dictionary<string, List<string>> Tabs { get; }
     public HashSet<string> Fields { get; }
     public HashSet<string> Grids { get; }
+}
+
+public class TabDetail
+{
+    public Dictionary<string, SectionDetail> Sections { get; } = [];
+}
+
+public class SectionDetail
+{
+    public HashSet<string> Controls { get; } = [];
 }